refactor: second cleanup pass, and partial indexes for the graded overlays
The five graded overlays were registered in four different shapes — three
standalone blocks and a two-entry loop — with the stage names spelled out again
in `key_layers` and a third time in the legend. `graded_dots::ALL` is now the
one table the pipeline iterates.

`GradedDots.grade` held either a column name or arbitrary SQL, which was the
renderer compensating for an imposm expression column yielding text. It becomes
`column` plus a `Grade` saying how to read it, and that turns out to matter for
more than tidiness: the filter is now a bare column predicate, so a partial
GiST index on it is usable. `COALESCE(NULLIF(x,'')::int,0) > 0` is not something
the planner can prove implies `x <> ''`, so it fell back to scanning every road
in the tile — measured on the real table, the two forms pick different plans.

`sql/additional.sql` grows a partial index per grade. These columns are tagged
on a small fraction of a million ways, so without them every overlay tile reads
the heap for all roads in its bbox.

The rest is one owner per fact: the quality range and the lossy-format set had
spread across `ImageFormat` and the export route, so `from_parts` takes the two
parts and `parse` is the token wrapper over it; `ExportLayer::TOGGLEABLE` was a
hand-kept list restating `RenderLayer::is_base`; the via ferrata ramp was a
byte-identical copy of the mtb one, now an alias, since both are alpine
difficulty read off a path and a reader moving between the overlays should not
have to relearn the colours.

+++ src/app/tile_variants.rs
@@ -94,7 +94,7 @@ impl FromStr for TileVariants {
                     add: HashSet::new(),
                     omit: HashSet::new(),
                 },
-                format: ImageFormat::Jpeg(DEFAULT_JPEG_QUALITY as u8),
+                format: ImageFormat::Jpeg(DEFAULT_JPEG_QUALITY),
                 tile_cache_base_path: None,
                 tile_index: None,
                 coverage_geojson: None,
@@ -176,7 +176,7 @@ mod tests {
         assert_eq!(v.len(), 3);
 
         assert_eq!(v[0].url_path, "/");
-        assert_eq!(v[0].format, ImageFormat::Jpeg(DEFAULT_JPEG_QUALITY as u8));
+        assert_eq!(v[0].format, ImageFormat::Jpeg(DEFAULT_JPEG_QUALITY));
         assert!(v[0].layers.base_map);
         assert!(v[0].layers.draws(RenderLayer::Shading));
         assert_eq!(
